+++ serde-version/src/version_groups/mod.rs
@@ -1,0 +1,6 @@
+mod macros;
+mod resolver;
+mod uri;
+
+pub use resolver::{DefaultVersionGroupResolver, VersionGroupResolver};
+pub use uri::{VersionGroupURI, VersionGroupURIs};